% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat_sheet_axis.R
\name{cheat_sheet_axis}
\alias{cheat_sheet_axis}
\title{Open a cheat sheet for the axis functions and formatting (base)}
\usage{
cheat_sheet_axis()
}
\value{
opens a pdf cheat sheet in the users' default browser
}
\description{
Open a cheat sheet for the axis functions and formatting (base)
}
\examples{
cheat_sheet_axis()
}
