% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat_sheet_lines_arrows_segments.R
\name{cheat_sheet_lines_arrows_segments}
\alias{cheat_sheet_lines_arrows_segments}
\title{Open a cheat sheet for the line, segments and arrows functions and their formatting (base)}
\usage{
cheat_sheet_lines_arrows_segments()
}
\value{
opens a pdf cheat sheet in the users' default browser
}
\description{
Open a cheat sheet for the line, segments and arrows functions and their formatting (base)
}
\examples{
cheat_sheet_lines_arrows_segments()
}
