% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat_sheet_open_all.R
\name{cheat_sheet_open_all}
\alias{cheat_sheet_open_all}
\title{Opens all cheat sheets currently implemented in the vocale package}
\usage{
cheat_sheet_open_all()
}
\value{
opens pdfs of each cheat sheet in the users' default browser
}
\description{
Opens all cheat sheets currently implemented in the vocale package
}
\examples{
cheat_sheet_open_all()
}
