% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat_sheet_plot.R
\name{cheat_sheet_plot}
\alias{cheat_sheet_plot}
\title{Open a cheat sheet for the plot function and its formatting (base)}
\usage{
cheat_sheet_plot()
}
\value{
opens a pdf cheat sheet in the users' default browser
}
\description{
Open a cheat sheet for the plot function and its formatting (base)
}
\examples{
cheat_sheet_plot()
}
