% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat_sheet_points_symbols_characters.R
\name{cheat_sheet_points_symbols_characters}
\alias{cheat_sheet_points_symbols_characters}
\title{Open a cheat sheet for the functions and formatting relating to plotting points, symbols and characters (base)}
\usage{
cheat_sheet_points_symbols_characters()
}
\value{
opens a pdf cheat sheet in the users' default browser
}
\description{
Open a cheat sheet for the functions and formatting relating to plotting points, symbols and characters (base)
}
\examples{
cheat_sheet_points_symbols_characters()
}
