% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cheat_sheet_title_mtext_text_subtitle_axis_labels.R
\name{cheat_sheet_title_mtext_text_subtitle_axis_labels}
\alias{cheat_sheet_title_mtext_text_subtitle_axis_labels}
\title{Open a cheat sheet for functions relating to text and their formatting (base, e.g. mtext, text, title and axis)}
\usage{
cheat_sheet_title_mtext_text_subtitle_axis_labels()
}
\value{
opens a pdf cheat sheet in the users' default browser
}
\description{
Open a cheat sheet for functions relating to text and their formatting (base, e.g. mtext, text, title and axis)
}
\examples{
cheat_sheet_title_mtext_text_subtitle_axis_labels()
}
